+++ src/redux/mission/actions/Actions.js
@@ -1,0 +1,7 @@
+export const FETCH_MISSIONS = 'FETCH_MISSIONS';
+export const JOIN_MISSION = 'JOIN_MISSION';
+
+export const joinMission = (payload) => ({
+  type: 'JOIN_MISSION',
+  payload,
+});
